Fix Warn4244 convert double to size_t... No longer a thing.

+++ Project/Model/IModel.hpp
@@ -33,7 +33,7 @@ public:
 
 	virtual void setTextureMap(Core::String _modOrigin, Core::String _textureMapName) = 0;
 
-	virtual void preRender(const double& update_ms);
+	virtual void preRender(const size_t& update_ms);
 
 	// Inherited via Drawable
 	virtual void draw(sf::RenderTarget & target, sf::RenderStates states) const override;
